export current state info in easy-to-digest way
If --export-dir is given, the state of specific things
will get exported as individual files in that directory,
and only if the parameter has changed since the last
time the state was exported for that key. This is intended
to make it easy for other lightweight software on the
same system to do, effectively, "if file X has changed,
take an action", generally via inotify or similar.

I *really* don't like this way of doing things, but it
will do for the very short term, at least.