adjustments to loop length handling
Loop lenght in bars is now stored to overall state, and
export files.

Log when the loop length is 0, as "reset"

If the previous loop length was 0 and now it's not-0, and
we're playing back, capture that timestamp as the start of
playback, since most likely we just finished laying down
the first layer of a loop.