ojbpm: figure out the Loopstation tempo and do stuff with it (#1)
A set of scripts for reading the MIDI data coming from a RC-202 Loopstation and do things with it (at the moment, "things" is "update the playback speed of an animation to match with the tempo of the current loop"). Should be in a fully functional state and relatively stable.