ojbpm: calculate loop len & post to discord
We previously calcualted a loop length in seconds, but never wrote it
out, because we were unsure how reliable it was. It seems to be pretty
reliable, though, *if* we assume 2/4 (which is what Jon normally uses),
so write that out now, and post a discord message about it, too.

It'll be wrong for non-2/4 things, of course. Not much we can do about
that, since the looper just doesn't give us that info.