looperstream: execute optional camera config script
...a few seconds after the camera is started. This feels ugly, but
best I can tell there's no way to have all the config values set via
ffmpeg, and it seems like opening the camera causes at least some of
them to reset, so we have to open it, and then actually run the v4l2-ctl
commands to set up the camera.

Also adds a `--force` option to force the camera to come up without
checking if the stream is actually running.