obsstats: use hardcoded list of possible outputs
When it comes to getting stats for OBS outputs, the ideal is that
we would query the list of outputs, and then query the status of
each in turn.

But, we had a recent OBS crash, where the stack indicated that
the crash was in the websocket code that enumerates outputs. So
to try to prevent this, we're going to just specify a list of
outputs that we want to explicitly query, and hope that avoids the
problematic enumeration.

I'm not convinced that there's not something else causing the crash,
under the hood, but this is easy enough to do.

FWIW, the relevent part of the stack of the crash was:

```
obs.dll!get_const_root+0x13
obs.dll!video_output_get_width+0xe
obs-websocket.dll!`Utils::Obs::ArrayHelper::GetOutputList'::`2'::<lambda_1>::operator()+0x597
obs.dll!obs_enum_outputs+0x48
obs-websocket.dll!Utils::Obs::ArrayHelper::GetOutputList+0x25
obs-websocket.dll!RequestHandler::GetOutputList+0x3e
obs-websocket.dll!RequestHandler::ProcessRequest+0x19e
obs-websocket.dll!WebSocketServer::ProcessMessage+0x14af
obs-websocket.dll!`WebSocketServer::onMessage'::`2'::<lambda_1>::operator()+0xc72
```