Add autohighlight feature for semiautomated short creation (#3)
Not quite at the quality we really want, but is relatively functional at this point, and gets the job done.

This is a series of scripts that is intended to be used something like:
- OBS (via advanced scene switcher) writes the string "HIGHLIGHT SEGMENT REQUEST" to the OBS log in response to a StreamDeck button press
- The `metaremux.py` script remuxes saved media (normal, clean feed, looper, stem recording), parses the starting timestamp (either by OCRing a burned-in timecode (best), or parsing the filename (imprecise)), and writes an adjunct "meta" file with timestamp information and a few other pieces of metadata
- The `ahllogparse.py` script searches through provided OBS logs and generates an index of timestamps for requested highlights
- The `ahlextract.py` script uses the outputs of the previous two scripts to extract short segments of the recordings (generally the "clean" recording) into small files containing 2 minutes of audio/video
- The `autocrop.py` script takes the extracted files, skips a few seconds in (see below), extracts a short amount of video (see below), does Jon detection on each frame, and crops to a 9:16 (vertical) aspect ratio with appropriate panning, nominally ready to post

Cropping times: Right now, the cropping skips 3 seconds into the video (to give Jon time to get away from the highlight button and back to playing) and then grabs 1 minute of video, the idea being that this will provide a probably-reasonable chunk of video that can be posted with near-zero amounts of effort. The uncropped highlights are a bit longer, so that if Jon (or whoever) wants to edit those prior to posting, there's a bit more footage to work with. This will probably be more configurable in the future.

There are probably still some issues with this whole process, but it does more or less actually work.

Right now the Jon detection runs on the CPU, so takes ~2x realtime on a fairly modern system. This is fine for now. For whatever reason, running on a GPU only speeds this up slightly. More optimization needs to be done here.