ongcodebot: Do (much) fuzzlier matching on song names (#8)
Now loads the title database into memory at startup, and does fuzzy
matching using the `rapidfuzz` python module. `fuzz.partial_ratio`
is our preferred scorer, which testing says is probably adequate for
these purposes, though not particularly great. This means the bot
can now find things that are even vaguely spelled correctly, with the
downside that results are far more numerous. Right now the cutoff
is a 60% match, though we might want to raise this threshold depending
on how the bot works in practice. We might want to make it configurable
on the command line, too.

The performance on this is surprisingly good -- most searches complete
in less than 5ms, which seems more than sufficient for our purposes.

This change also adds the `--benchmark` command-line option (to test
performance), and the `--search` command-line option (to perform a
search, just as would happen if using the command from discord)